requirements.txt
requests
bs4
streamlit
openai
faiss-cpu
python-dotenv
pdfplumber
numpy

